Fix multiple failures on a single worker
